% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{calc.select2}
\alias{calc.select2}
\title{Utility function2}
\usage{
calc.select2(m)
}
\arguments{
\item{m}{a vector of 4 elements}
}
\value{
A-selection rate
}
\description{
Utility2
}
\examples{
calc.select2(c(1,2,3,4))
}
\keyword{NA}

