% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{calc.success2}
\alias{calc.success2}
\title{Utility function}
\usage{
calc.success2(m)
}
\arguments{
\item{m}{a vector of 4 elements}
}
\value{
success rate
}
\description{
Utility
}
\examples{
calc.success2(c(1,2,3,4))
}
\keyword{NA}

