% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{dist.success.select}
\alias{dist.success.select}
\title{2D-grid probability of Success rate, SelectionA rate}
\usage{
dist.success.select(os, prob.v)
}
\arguments{
\item{outerseries}{output}

\item{tab.prob.unlisted}{}
}
\value{
a list of matrix (x axis is success rate and y axis is selection.A rate)
}
\description{
2D-grid probability of Success rate, SelectionA rate
}
\examples{
#### Step 1
N <- 20
out <- serialTables(N)
outseries <- unlistSerialTables(out)

#### Step 2
prob.select.exp <- prob.expect.select(out[[2]])
targets <- seq(from=0,to=1,by=0.1)
prob.select.selfs <- list()
for(i in 1:length(targets)){
 prob.select.selfs[[i]] <- prob.self.select(out[[2]],targets[i])
}
prob.select.list <- prob.select.selfs
prob.select.list[[length(prob.select.list)+1]] <- prob.select.exp

#### Step 3
pop1 <- c(0,0,0,0,0,0,0,0,0,0,0,1)
select1 <- prob.select.mix(prob.select.list,pop1)
p.AB <- c(0.8,0.5)
tab.prob <- table.prob(select1,p.AB)
tab.prob.unlisted <- lapply(tab.prob,unlist)
dist.ss.out <- dist.success.select(outseries,tab.prob.unlisted)
for(i in 1:length(dist.ss.out)){
	persp(dist.ss.out[[2]][[i]],theta=80,phi=50,shade=0.75)
}
}
\keyword{NA}

