% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{prob.expect.select}
\alias{prob.expect.select}
\title{Make a list of expect-dependent selection probability}
\usage{
prob.expect.select(tablist)
}
\arguments{
\item{tablist}{list of matrices with 4 columns}
}
\value{
list of selection probability
}
\description{
Make a list of expect-dependent selection probability
}
\examples{
N <- 20
out <- serialTables(N)
prob.select.exp <- prob.expect.select(out[[2]])
}
\keyword{NA}

