% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{prob.expected.3}
\alias{prob.expected.3}
\title{Expected value of beta-distribution-dependent decision}
\usage{
prob.expected.3(x)
}
\arguments{
\item{x}{a vector of 4 elements}
}
\value{
probability to select A
}
\description{
Expected value of beta-distribution-dependent decision
}
\examples{
prob.expected.3(c(1,2,3,4))
}
\keyword{NA}

