% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{prob.list}
\alias{prob.list}
\title{Make a list of selection probability}
\usage{
prob.list(tablist, f)
}
\arguments{
\item{tablist}{list of matrices with 4 columns}

\item{f}{function to calculate decision probability}
}
\value{
list of vectors
}
\description{
Make a list of selection probability
}
\examples{
N <- 20
out <- serialTables(N)
prob.success.v <- prob.list(out[[2]],calc.success2)
prob.select.A <- prob.list(out[[2]],calc.select2)
}
\keyword{NA}

