% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{prob.select.mix}
\alias{prob.select.mix}
\title{Make a list of selection probability for a mixture of selection types}
\usage{
prob.select.mix(L, v)
}
\arguments{
\item{list}{of selection probabilities}

\item{a}{weight vector}
}
\value{
list of selection probability
}
\description{
Make a list of selection probability for a mixture of selection types
}
\examples{
N <- 20
out <- serialTables(N)
prob.select.exp <- prob.expect.select(out[[2]])
targets <- seq(from=0,to=1,by=0.1)
prob.select.selfs <- list()
for(i in 1:length(targets)){
 prob.select.selfs[[i]] <- prob.self.select(out[[2]],targets[i])
}
prob.selection.list <- prob.select.selfs
prob.selection.list[[length(prob.selection.list)+1]] <- prob.select.exp
weight.v <- runif(length(prob.selection.list))
weight.v <- weight.v/sum(weight.v)
prob.selection.this <- prob.select.mix(prob.selection.list,weight.v)
}
\keyword{NA}

