% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{prob.self.3}
\alias{prob.self.3}
\title{Probability of better than target success rate-dependent self decision}
\usage{
prob.self.3(x, target)
}
\arguments{
\item{x}{a vector of 4 elements}
}
\value{
probability to select A
}
\description{
Probability of better than target success rate-dependent self decision
}
\examples{
prob.self.3(c(1,2,3,4))
}
\keyword{NA}

