% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{prob.self.select}
\alias{prob.self.select}
\title{Make a list of self-decision selection probability}
\usage{
prob.self.select(tablist, target)
}
\arguments{
\item{tablist}{list of matrices with 4 columns}

\item{target}{a value from 0 to 1}
}
\value{
list of selection probability
}
\description{
Make a list of self-decision selection probability
}
\examples{
N <- 20
out <- serialTables(N)
targets <- seq(from=0,to=1,by=0.1)
prob.select.selfs <- list()
for(i in 1:length(targets)){
 prob.select.selfs[[i]] <- prob.self.select(out[[2]],targets[i])
}
}
\keyword{NA}

