% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{relist.sum}
\alias{relist.sum}
\title{Utility function to handle lists}
\usage{
\method{relist}{sum}(L1, L2)
}
\arguments{
\item{L1}{list}

\item{L2}{list with the same form as L1}
}
\value{
list
}
\description{
Utility function to handle lists
}
\examples{
L1 <- list(c(1,2,3),matrix(1:4,2,2))
L2 <- list(c(2,3,4),matrix(2:5,2,2))
relist.sum(L1,L2)
}
\keyword{NA}

