% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{table.prob}
\alias{table.prob}
\title{Probability of all tables for particular selection pattern and success rate fo A and B arms}
\usage{
table.prob(s1, p)
}
\arguments{
\item{list}{of A-selection probability}

\item{a}{vector of two arms success rate}
}
\value{
list of list of probability of tables
}
\description{
Probability of all tables for particular selection pattern and success rate fo A and B arms
}
\examples{
#### Step 1
N <- 20
out <- serialTables(N)
outseries <- unlistSerialTables(out)

#### Step 2
prob.select.exp <- prob.expect.select(out[[2]])
targets <- seq(from=0,to=1,by=0.1)
prob.select.selfs <- list()
for(i in 1:length(targets)){
 prob.select.selfs[[i]] <- prob.self.select(out[[2]],targets[i])
}
prob.select.list <- prob.select.selfs
prob.select.list[[length(prob.select.list)+1]] <- prob.select.exp

#### Step 3
pop1 <- c(0,0,0,0,0,0,0,0,0,0,0,1)
select1 <- prob.select.mix(prob.select.list,pop1)
p.AB <- c(0.8,0.5)
tab.prob <- table.prob(select1,p.AB)
}
\keyword{NA}

