% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/decision2x2.R
\name{unlistSerialTables}
\alias{unlistSerialTables}
\title{From the output of serialTables(), multiple lists are generated}
\usage{
unlistSerialTables(out)
}
\arguments{
\item{out;}{serialTables() ' out}
}
\value{
list of success rate for tables

list of selection A rate for tables

list of A&Success, list of A&Failure, list of B&Success and list of B&Failure
}
\description{
From the output of serialTables(), multiple lists are generated
}
\examples{
N <- 20
out <- serialTables(N)
outseries <- unlistSerialTables(out)
}
\keyword{NA}

